python/requirements.txt
firebase_admin==3.0.1
requests
python_dotenv